Serialize ValueStringBuilder allocation tests with [DoNotParallelize]
The two #if !DEBUG MemoryWatch tests assert zero allocations via GC.GetAllocatedBytesForCurrentThread(). Under the assembly's method-level parallelism a concurrent test can drain a shared ArrayPool bucket between the warmup and the measured rent, turning a pool-satisfied rent into a real allocation and failing the watch intermittently in Release. Extract them into a [DoNotParallelize] ValueStringBuilderTests_Memory class, mirroring DefaultInterpolatedStringHandlerTests_Memory.
